[imagebuilder] don't passthru AWS creds when using IAM role
By default `imagebuilder` will pass through AWS creds to the target EC2
instance.

When the instance is using an IAM role (specified by `InstanceProfile`
in config) this almost certainly isn't what the user wants, so stop
doing it.

Note that this *can* be worked around already by unexporting the
`AWS_*` environment variables in the `bootstrap-vz` template, but this
is pretty unintuitive. Particularly when a user debugging a failing
build might SSH to their `imagebuilder` instance and be surprised to
find that things failing in their build process work just fine when
run on the commandline from the same instance...

Also, while we're in there, when *not* using IAM role, export
`AWS_SECURITY_TOKEN` with the same value as `AWS_SESSION_TOKEN` for
improved older AWS client compatibility.